core/deps-tools.mk: New target `list-deps` to list deps recurvively
Therefore, it lists direct dependencies, dependencies' dependencies, and
so on.

The following targets are also available to list other kinds of
dependencies:
    make list-doc-deps
    make list-rel-deps
    make list-test-deps
    make list-shell-deps

In all cases, they list "normal" and eg. "test" dependencies of the
top-level project, then list only "normal" dependencies' dependencies.

It's possible to list several kinds in one go:
    make list-deps DEP_TYPES='doc test'

As it may be difficult to use the output of these targets because
the list could appear after other targets output (like dependencies
fetching), the list is available in files pointed by the following
variables:
    $(ERLANG_MK_RECURSIVE_DEPS_LIST)
    $(ERLANG_MK_RECURSIVE_DOC_DEPS_LIST)
    $(ERLANG_MK_RECURSIVE_REL_DEPS_LIST)
    $(ERLANG_MK_RECURSIVE_TEST_DEPS_LIST)
    $(ERLANG_MK_RECURSIVE_SHELL_DEPS_LIST)